Fix UI not removing bonus when set to 0
Previously, when the player reset the bonus of a player (human or AI)
back to zero, the UI maintains the last selected bonus. This should fix
that bug.

The reason for the bug was that the `UpdateUserBattleStatus` function
used to override the visibility status of the bonus (handicap) label
based on whether the user is playing or spectating. To elaborate, when
the bonus is set to zero, the function sets the visibility of the label
to false so nothing shows up. But then later in the code, the function
mistakenly reshows the label if the user is playing.