configuration.lua - added key enableInspector (default = false)
dbg_chili_inspector.lua
- set conf value enableInspector to false onClose

gui_settings_window.lua

AddCheckboxSettings(..)
- use Object:onClick instead of Checkbox:onChange

Why onClick:
Because Checkbox:onChange Listeners are called BEFORE .checked of Checkbox is set to new value. (This feels totally wrong, but could be handled on another PR!)

If we want to ask for .checked property in a sub sub function of onchange(), it's still on old value. Bad.
I changed this here, because chili_inspector can be closed by 2 actors:

1. dev-settings
2. close button of inspector itself

So, when inspector:close changes conf value to false, it triggers onConfigurationChange, which then must change the checked prop of settings window.
But changing the checked value triggers onChange , which leads in re-checking it again.