sort teams in battle window by team number and remove almost all empty teams and fix update battlestatus for bots
When changing from ffa/16 to team8v8 a lot of teams stayed empty and open, because only the team with highest teamIndex was removed.

Now the only possible empty teams are those of teamIndex 1 and 2.
They could be removed too when more teams exists , but it's better for new players to always have team 1 and 2 existing, especially in Singleplayer. Otherwise they can easily end up with e.g. battle window showing team 3 and 5 and they would need to adjust the startboxes accordingly.

In multiplayer with spads autobalance turned on there won't be an empty team of teamID= 0 or 1. Without autobalance it's possible, but users can just fill up or keep it empty.