show running time and last ended time in tooltip
1. Chobby now actively stores timestamps when a server goes from lobby_state to ingame_state locally (which is read from CLIENTSTATUS-changes of lobby founder)
2. If the last switch of a server was before chobby login, spads is asked by RPC commands for the current running time or last game ended time ONCE
For absolutely prohibiting spads command spam, any request to spads is stored and therefore each host instance can only be asked once in the lifetime of chobby at all.

The requests to spads are only done on user needs. That is, if a user hovers over a host, where times are still unknown.
IF a user is logged in for a longer time, there is no need to send commands to spads, because chobby registered starts and stops of servers already.

A fresh logged in chobby with a user hovering his mouse pointer over the whole battlelist may easily produce 20-30 commands to 20-30 spads hosts.
But only one command ever in the whole login session for each host.