fix: show current playercount in battle list
At OnJoinBattle and OnLeftBattle the widget's internal var previousPlayerCount wasn't updated, while it was done in UpdateBattleInfo. Leading to inconsistent values in previousPlayerCount.

E.G.
As a result OnJoinedBattle could increase the shown playercount from 16 to 17 (which usually happens for a seconds before spads forces spec), but when the 17th. player is auto-switched to spec, updatebattleInfo saw the old previousPlayerCount of 16, which is same as the received new one and didn't update shown number to 16 and we saw 17 in battle list window.