disabled local generated game times
As long as incoming server messages do not get stored with a timestamp, we can't use their executing time as a reliable source of truth for game times.
The main issue is the delay while ingame until they get unbuffered. We don't want to use the timestamp of unbuffering for other battles start or end time.

This may get enabled again after implementing timestamp storing for server commands.