modoptions: add `teamffa_start_boxes_shuffle`
In 79a5ecf276cbbfafb2680d4485da7c9503b0b05d on the game repository (see
[1]), we reworked FFA start points and additionally added a shuffling
system for start boxes in TeamFFA games.

This was added because in multiplayer games it is desirable to hide
start location, however in singleplayer it is a common scenario to play
a pseudo-TeamFFA game against multiple AIs, and in this case players
typically do not want to be booted off their specifically chosen start
box.

To allow for toggling this behaviour on or off as needed, we add a new
ModOption in the `Extra` tab (next to `Anonymous Mode` since it is
another common setting tweaked in FFA games). This ModOption is set to
"on" by default such that behaviour is unchanged in multiplayer games,
however we also introduce a small QoL where it defaults to "off" in
singleplayer games via a custom override done in Chobby.

[1]: https://github.com/beyond-all-reason/Beyond-All-Reason/commit/79a5ecf276cbbfafb2680d4485da7c9503b0b05d