Add a "Muted" status icon for players listed in a lobby.
This patch allows Chobby to understand BarManager updates which
include a "mutes" section, and to display a status icon next to
players who are referenced in that section (when they are in the
lobby).

A corresponding patch to BarManager is required, so that it
actually includes a "mutes" section in its updates. If BarManager
has not been updated, then this patch will have no visible effect.