feat(Change Team): use drop down window
Replace the integer selection window for changing teams (a slider) with
a combobox (a drop down box) for selecting teams. This should feel more
intuitive and be easier to click through. This also matches the SPADS
configuration interface where the number of teams and team size are
organised with a combobox.