Add rejoin battle button
Add a button that allows users to rejoin a battle if they did not properly leave a battle, i.e. press leave battle, or leave lobby.

This allows users to reconnect to lobbys that are open after a break or to allow users to rejoin a server in the event of a crash. No longer do people have to keep a mental note of what game they were in.