Add rejoin battle button (#1034)
Add a button that allows users to rejoin a battle if they did not
properly leave a battle, i.e. press leave battle, or leave lobby.

This allows users to reconnect to lobbies that are open after a break or
to allow users to rejoin a server in the event of a crash. No longer do
people have to keep a mental note of what game they were in.

To note, a known bug exists wherein if a BattleLobby opens up with the
same the same ID as the last BattleLobby when it was closed, the user
will connect to that 'random' lobby. This bug ultimately is very small
though.

---------

Co-authored-by: Fireball <32598847+FIr3baL@users.noreply.github.com>