Update json.lua, deprecate duplicate json.lua scripts, update callsites to new version. (#1051)
Third time's the charm. This PR should hopefully be much easier to
review thanks to the teaching efforts of fellow contributors.

Json.lua is being updated to the same version that is being used in the
BAR game repo.
A VFS.Include() function is being added into LuaMenu's main.lua script
so that the json.lua script only then needs to be called in once.
Duplicate json.lua scripts were deprecated.
Finally, we edited effected modules to utilize the updated json.lua
script.

This is work continued from #1050 .