Fix hang/crash when renaming a long battle title (#1077)
Closes #1002 

The UTF-8 previous character parsing function in the unicode.lua header
used in chili seems to have have infinite loop vulnerabilities when
dealing with edge cases in long strings. This would result in hangs or
crashes when selecting text and backspacing while editing long titles in
the Rename Battle dialog.