Feature/bar/5694 reduce default audio volume (#1082)
Closes
https://github.com/beyond-all-reason/Beyond-All-Reason/issues/5694

Lower BAR default master volume from 80 (technically 100 from the config
file but it gets mathed down to 80) to 30

Lower BYAR-Chobby default Menu Music Volume from 0.5 to 0.3

This helps with new players not getting their ears blasted, which has
happened to some of my friends and the person in the issue linked above
(and their friend(s))

# Testing

To test this, run

```
find -name 'springsettings*.cfg' -exec bash -c 'mv "${0}" "${0}.bak"' {} \;
```
and
```
find -name 'IGL_data.lua' -exec bash -c 'mv "${0}" "${0}.bak"' {} \;
```
in the beyond all reason root directory... or in Windows.... uhhhhh,
switch to linux?

Then open bar launcher, set to dev lobby->start, then settings to see
Menu Music Volume is 0.3 instead of 0.5, then
Skirmish->Start->...waitforload...->Settings->Audio to see Master Volume
is 30 instead of 80 or whatever.

Better way to test it is to get a new bar install, but that would take a
lot of time and effort, and I'm pretty sure this works.

I have a BAR branch that changes all the various default values (for if
some error occurs in pulling it from the config file), so if you want me
to make a PR for that, lmk.